@@ -1,4 +1,4 @@
-class AddColumnsSalaryAndLevelToUsers < ActiveRecord::Migration
+class AddColumnsSalaryAndLevelToUsers < ActiveRecord::Migration[4.2]
   def change
     add_column :users, :salary, :decimal, scale: 2, precision: 100
     add_column :users, :level, :decimal, scale: 2, precision: 100
