@@ -1,4 +1,4 @@
-class AddLevelOnlyStatusToReviews < ActiveRecord::Migration
+class AddLevelOnlyStatusToReviews < ActiveRecord::Migration[4.2]
   def change
     add_column :reviews, :is_level_only, :boolean, default: false
   end
