@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -11,67 +10,64 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20180310063710) do
+ActiveRecord::Schema.define(version: 2018_03_10_063710) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
 
-  create_table "comments", force: :cascade do |t|
-    t.integer  "commentable_id"
-    t.string   "commentable_type"
-    t.string   "title"
-    t.text     "body"
-    t.string   "subject"
-    t.integer  "user_id",          null: false
-    t.integer  "parent_id"
-    t.integer  "lft"
-    t.integer  "rgt"
+  create_table "comments", id: :serial, force: :cascade do |t|
+    t.integer "commentable_id"
+    t.string "commentable_type"
+    t.string "title"
+    t.text "body"
+    t.string "subject"
+    t.integer "user_id", null: false
+    t.integer "parent_id"
+    t.integer "lft"
+    t.integer "rgt"
+    t.datetime "created_at"
+    t.datetime "updated_at"
+    t.index ["commentable_id", "commentable_type"], name: "index_comments_on_commentable_id_and_commentable_type"
+    t.index ["user_id"], name: "index_comments_on_user_id"
+  end
+
+  create_table "review_years", id: :serial, force: :cascade do |t|
+    t.text "name"
     t.datetime "created_at"
     t.datetime "updated_at"
   end
 
-  add_index "comments", ["commentable_id", "commentable_type"], name: "index_comments_on_commentable_id_and_commentable_type", using: :btree
-  add_index "comments", ["user_id"], name: "index_comments_on_user_id", using: :btree
-
-  create_table "review_years", force: :cascade do |t|
-    t.text     "name"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+  create_table "reviews", id: :serial, force: :cascade do |t|
+    t.integer "status", default: 0
+    t.integer "reviewer_id"
+    t.integer "reviewee_id"
+    t.text "text"
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.decimal "suggested_level", precision: 100, scale: 2
+    t.integer "review_year_id"
+    t.boolean "is_level_only", default: false
+    t.index ["reviewee_id"], name: "index_reviews_on_reviewee_id"
+    t.index ["reviewer_id"], name: "index_reviews_on_reviewer_id"
   end
 
-  create_table "reviews", force: :cascade do |t|
-    t.integer  "status",                                    default: 0
-    t.integer  "reviewer_id"
-    t.integer  "reviewee_id"
-    t.text     "text"
-    t.datetime "created_at",                                                null: false
-    t.datetime "updated_at",                                                null: false
-    t.decimal  "suggested_level", precision: 100, scale: 2
-    t.integer  "review_year_id"
-    t.boolean  "is_level_only",                             default: false
-  end
-
-  add_index "reviews", ["reviewee_id"], name: "index_reviews_on_reviewee_id", using: :btree
-  add_index "reviews", ["reviewer_id"], name: "index_reviews_on_reviewer_id", using: :btree
-
-  create_table "users", force: :cascade do |t|
-    t.string   "email",                                         default: "", null: false
-    t.string   "name"
-    t.string   "image"
+  create_table "users", id: :serial, force: :cascade do |t|
+    t.string "email", default: "", null: false
+    t.string "name"
+    t.string "image"
     t.datetime "remember_created_at"
-    t.integer  "sign_in_count",                                 default: 0,  null: false
+    t.integer "sign_in_count", default: 0, null: false
     t.datetime "current_sign_in_at"
     t.datetime "last_sign_in_at"
-    t.string   "current_sign_in_ip"
-    t.string   "last_sign_in_ip"
+    t.string "current_sign_in_ip"
+    t.string "last_sign_in_ip"
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.decimal  "salary",              precision: 100, scale: 2
-    t.decimal  "level",               precision: 100, scale: 2
-    t.decimal  "decided_level",       precision: 100, scale: 2
+    t.decimal "salary", precision: 100, scale: 2
+    t.decimal "level", precision: 100, scale: 2
+    t.decimal "decided_level", precision: 100, scale: 2
+    t.index ["email"], name: "index_users_on_email", unique: true
   end
-
-  add_index "users", ["email"], name: "index_users_on_email", unique: true, using: :btree
 
   add_foreign_key "reviews", "review_years"
   add_foreign_key "reviews", "users", column: "reviewee_id"
